sagittarius: hook: Allow installing git hooks in repositories
Add a new 'hooks' field to the repository configuration. This member
should be an association list where each key is a hook name and each
value is a thunk to run for that hook.

On reconfigure, sagittarius will install a hook that will load and
call the thunk from the config file.

The hooks are executed via a new CLI command 'hook' that takes the
repository name and the hook name as arguments.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>